+++ hf-app/database/migrations/2021_02_23_203413_create_user_parameters_table.php
@@ -19,8 +19,8 @@ class CreateUserParametersTable extends Migration
             $table->foreignId('parameter_id')->constrained();
             $table->integer('measurement_times')->nullable();
             $table->enum('measurement_span', ['hour', 'day', 'week', 'month'])->nullable();
-            $table->integer('threshold_min')->nullable();
-            $table->integer('threshold_max')->nullable();
+            $table->float('threshold_min')->nullable();
+            $table->float('threshold_max')->nullable();
             $table->timestamps();
         });
     }
